% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{boundAngle}
\alias{boundAngle}
\title{Convert an angle in radians into a standard form.
Takes any angle in radians and returns an angle beween pi and -pi.}
\usage{
boundAngle(theta)
}
\arguments{
\item{Theta}{an angle in radians}
}
\value{
an angle in radians between pi and -pi
}
\description{
Convert an angle in radians into a standard form.
Takes any angle in radians and returns an angle beween pi and -pi.
}
