% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{computeTangents}
\alias{computeTangents}
\title{Turn a shape into a set of tangents as offsets from the unit
circle.}
\usage{
computeTangents(x, y)
}
\arguments{
\item{x}{The x coordinates for the points which make up the digitised curve}

\item{y}{The y coordinates for the points which make up the digitised curve}
}
\value{
Tangents
}
\description{
Turn a shape into a set of tangents as offsets from the unit
circle.
}
