% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{distance}
\alias{distance}
\title{Compute Euclidean distance between two points}
\usage{
distance(x1, y1, x2, y2)
}
\arguments{
\item{two}{points to find the distance between}
}
\value{
Euclidean distance between the two points
}
\description{
Compute Euclidean distance between two points
}
