% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spline.R
\docType{data}
\name{estimateDeli}
\alias{estimateDeli}
\title{Numerically estimate the value of deli for a curve segment of
the spline.}
\format{\preformatted{ NULL
}}
\usage{
estimateDeli
}
\description{
Numerically estimate the value of deli for a curve segment of
the spline.
}
\keyword{datasets}

