% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcurve.R
\name{fftoToCoeffs}
\alias{fftoToCoeffs}
\title{Convert fourier coefficients to Ak, sk values.}
\usage{
fftoToCoeffs(fftCoeffs)
}
\arguments{
\item{Coeffs}{The output of fft}
}
\value{
Ak and sk values for the reverse fourier
}
\description{
Convert fourier coefficients to Ak, sk values.
}
