% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{findCenter}
\alias{findCenter}
\title{Finds the center of a circle given three points
Equations taken from http://www.ambrsoft.com/TrigoCalc/Circle3D.htm}
\usage{
findCenter(x1, y1, x2, y2, x3, y3)
}
\arguments{
\item{points}{that make up the three points on the circumference}
}
\value{
a vector [x,y] that is the center of the circle
}
\description{
Finds the center of a circle given three points
Equations taken from http://www.ambrsoft.com/TrigoCalc/Circle3D.htm
}
