% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spline.R
\name{findTangentAngle}
\alias{findTangentAngle}
\title{Find that tangent angle to a point on a circle given it's center and
a point on its circumference}
\usage{
findTangentAngle(xc, yc, x, y)
}
\arguments{
\item{xc,}{yc coordinates of the circle's center}

\item{x,y}{coordinates of a point on the cicle's circumference}
}
\value{
tangent angle to the circle
}
\description{
Find that tangent angle to a point on a circle given it's center and
a point on its circumference
}

