% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\docType{data}
\name{fitSpline}
\alias{fitSpline}
\title{Fits a spline to the set of points provided.}
\format{An object of class \code{NULL} of length 0.}
\usage{
fitSpline
}
\arguments{
\item{points}{The points which make up the digitised curve}
}
\value{
Coeffiecients for the spline
}
\description{
Fits a spline to the set of points provided.
}
\keyword{datasets}
