% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{resampleTangents}
\alias{resampleTangents}
\title{Resample the tangents curve to a given number of samples.
Uses the Hangle spline function. This function has a lot of 
numerical approximation in it. You should always compare its
output to the original to make sure nothing when wrong.}
\usage{
resampleTangents(x, y, tangents, length.out = 1024)
}
\arguments{
\item{x}{The x coordinates for the points which make up the digitised curve}

\item{y}{The y coordinates for the points which make up the digitised curve}

\item{tangents}{A vector of tangents}

\item{length.out}{The length of the output array}
}
\value{
A resampled tangent vector with length length.out
}
\description{
Resample the tangents curve to a given number of samples.
Uses the Hangle spline function. This function has a lot of 
numerical approximation in it. You should always compare its
output to the original to make sure nothing when wrong.
}
