% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{sequenceAngles}
\alias{sequenceAngles}
\title{Make a sequence of angles in radians all follow each other in with a value
as close to the previous as possible}
\usage{
sequenceAngles(angles)
}
\arguments{
\item{angles}{An ordered vector of angles (in radians)}
}
\value{
a vector of angles where the adjacent angles are as close to each
 other in magnitude as possible
}
\description{
Make a sequence of angles in radians all follow each other in with a value
as close to the previous as possible
}
