% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{simpsonsRuleCell}
\alias{simpsonsRuleCell}
\title{Numerical integration of a single cell using
Simpson's rule.}
\usage{
simpsonsRuleCell(f, a, b)
}
\arguments{
\item{f}{The function to integrate. Must take exactly 1 parameter}

\item{a}{Lower region bound}

\item{b}{Upper region bound}
}
\value{
Numerical approximation of the integral of f on [a,b]
}
\description{
Numerical integration of a single cell using
Simpson's rule.
}
