% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcurve.R
\name{tangentsToXY}
\alias{tangentsToXY}
\title{Convert a list of curve tangents to xy points for plotting.
Each step along the curve is assumed to be equally spaced.}
\usage{
tangentsToXY(tangents)
}
\arguments{
\item{tangents}{Equally spaced tangents along the curve}
}
\value{
A list contaning a vector x points and a vector of y points.
}
\description{
Convert a list of curve tangents to xy points for plotting.
Each step along the curve is assumed to be equally spaced.
}
